+++ recipes/react-template/src/App.tsx
@@ -9,7 +9,7 @@ function App() {
       <header className="App-header">
         <Logo className="App-logo" />
         <p>
-          Edit <code>src/App.js</code> and save to reload.
+          Edit <code>src/App.tsx</code> and run <code>vr build</code>.
         </p>
         <a
           className="App-link"